api/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "rusticai-api"
version = "0.0.1"
description = "API Server for Rustic AI"
authors = ["Dragonscale Industries Inc. <dev@dragonscale.ai>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://www.rustic.ai/"
repository = "https://github.com/rustic-ai/python-framework"
packages = [{ include = "rustic_ai", from = "src" }]

[tool.poetry.urls]
"Rustic AI Core" = "https://pypi.org/project/rusticai-core/"

[tool.poetry.dependencies]
python = "^3.12"
rusticai-core = { path = "../core", develop = true }
rusticai-ray = { path = "../ray", develop = true }
fastapi = "^0.115.8"
starlette = "^0.45.3"
uvicorn = "^0.34.0"
ptvsd = "^4.3.2"
python-multipart = "^0.0.20"
opentelemetry-distro = "^0.52b1"
websockets = "^15.0.1"
griffe = "^1.7.2"
psycopg = "^3.2.6"

[tool.poetry-monorepo.deps]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = { version = "^24.2.0", extras = ["jupyter"] }
flake8 = "^7.1.2"
tox = "^4.24.1"
isort = "^6.0.0"
mypy = "^1.15.0"
pytest-asyncio = "^0.25.3"
httpx = "^0.28.1"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
plugins = "pydantic.mypy"
explicit_package_bases = true


[tool.isort]
profile = "black"
known_third_party = ["pydantic"]
known_first_party = ["rustic_ai.core", "rustic_ai.ray", "rustic_ai.api_server"]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope="function"
